+++ src/test/java/gift/service/OptionServiceTest.java
@@ -42,7 +42,7 @@ public class OptionServiceTest {
         // given
         ArgumentCaptor<Option> captor_o = ArgumentCaptor.forClass(Option.class);
 
-        OptionRequest optionRequest = new OptionRequest("[1] 기본", 1L);
+        OptionRequest optionRequest = new OptionRequest("[1] 기본", 1);
         Product product = new Product("product", 10000, "url", new Category("신규"));
         TestUtil.setId(product, 1L);
         given(productRepository.findById(1L)).willAnswer(invocation -> Optional.of(product));
@@ -62,7 +62,7 @@ public class OptionServiceTest {
         Product product = new Product("product", 10000, "url", new Category("신규"));
         for(int i = 1; i <= 5; i++){
             product.addOption(
-                    new Option("옵션" + i, (long)i)
+                    new Option("옵션" + i, i)
             );
         }
         TestUtil.setId(product, 1L);
@@ -83,8 +83,8 @@ public class OptionServiceTest {
     @DisplayName("updateTest")
     void test4(){
         // given
-        OptionRequest optionRequest = new OptionRequest("[1] 수정", 50L);
-        Option option = new Option("[1] 기본", 100L);
+        OptionRequest optionRequest = new OptionRequest("[1] 수정", 50);
+        Option option = new Option("[1] 기본", 100);
         TestUtil.setId(option, 1L);
         given(optionRepository.findById(1L)).willAnswer(invocation -> Optional.of(option));
         // when
@@ -100,8 +100,8 @@ public class OptionServiceTest {
         // given
         Product product = new Product("product", 10000, "url", new Category("신규"));
         TestUtil.setId(product, 1L);
-        Option option1 = new Option("[1] 옵션 1", 500L);
-        Option option2 = new Option("[2] 옵션 2", 500L);
+        Option option1 = new Option("[1] 옵션 1", 500);
+        Option option2 = new Option("[2] 옵션 2", 500);
         TestUtil.setId(option1, 1L);
         TestUtil.setId(option2, 1L);
         product.addOption(option1);
@@ -120,7 +120,7 @@ public class OptionServiceTest {
         // given
         Product product = new Product("product", 10000, "url", new Category("신규"));
         TestUtil.setId(product, 1L);
-        Option option1 = new Option("[1] 옵션 1", 500L);
+        Option option1 = new Option("[1] 옵션 1", 500);
         TestUtil.setId(option1, 1L);
         product.addOption(option1);
